// apps/frontend/app/lib/api/generated/api.types.ts
export interface components {
  schemas: {
// ...
    Document: {
      /** Format: uuid */
      id: string;
      title: string;
      description: string;
      type: string;
      sections: number;
      lastUpdated: string;
      storageUrl: string;
      downloadCount: number;
      actName: string | null;
      jurisdiction: string | null;
      sourceUrl: string | null;
      version: number;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
    };
// ...
    DocumentResource: {
      /** @enum {string} */
      type: "document";
      /** Format: uuid */
      id: string;
      /**
       * Document
       * @description Document response schema
       */
      attributes: {
        /** Format: uuid */
        id: string;
        title: string;
        description: string;
        type: string;
        sections: number;
        lastUpdated: string;
        storageUrl: string;
        downloadCount: number;
        actName: string | null;
        jurisdiction: string | null;
        sourceUrl: string | null;
        version: number;
        /** Format: date-time */
        createdAt: string;
        /** Format: date-time */
        updatedAt: string;
      };
      relationships?: {
        [key: string]: unknown;
      };
      meta?: {
        [key: string]: unknown;
      };
      links?: {
        [key: string]: string;
      };
    };
    DocumentSingleResponse: {
      data: {
        /** @enum {string} */
        type: "document";
        /** Format: uuid */
        id: string;
        /**
         * Document
         * @description Document response schema
         */
        attributes: {
          /** Format: uuid */
          id: string;
          title: string;
          description: string;
          type: string;
          sections: number;
          lastUpdated: string;
          storageUrl: string;
          downloadCount: number;
          actName: string | null;
          jurisdiction: string | null;
          sourceUrl: string | null;
          version: number;
          /** Format: date-time */
          createdAt: string;
          /** Format: date-time */
          updatedAt: string;
        };
        relationships?: {
          [key: string]: unknown;
        };
        meta?: {
          [key: string]: unknown;
        };
        links?: {
          [key: string]: string;
        };
      };
      links: {
        self: string;
      };
      metadata: {
        requestId: string;
        timestamp: string;
      } & {
        [key: string]: unknown;
      };
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}